+++ lib.php
@@ -36,6 +36,7 @@
 
 use mod_collaborate\local\collaborate_editor;
 use mod_collaborate\local\submissions;
+use core\url;
 
 /* Moodle core API */
 
@@ -562,9 +563,13 @@ function collaborate_extend_navigation(navigation_node $navref, stdClass $course
  * This function is called when the context for the page is a collaborate module. This is not called by AJAX
  * so it is safe to rely on the $PAGE.
  *
- * @param settings_navigation $settingsnav complete settings navigation tree.
- * @param navigation_node $collaboratenode collaborate administration node.
+ * @param settings_navigation $settingsnav Complete settings navigation tree.
+ * @param navigation_node $collaboratenode Collaborate administration node.
  */
 function collaborate_extend_settings_navigation(settings_navigation $settingsnav, ?navigation_node $collaboratenode=null) {
-    // Todo: Delete this function and its docblock, or implement it.
+    global $PAGE;
+
+    // Extend the settings nav with the namechanger page url.
+    $namechangeurl = new url('/mod/collaborate/namechanger.php', ['courseid' => $PAGE->course->id]);
+    $collaboratenode->add(get_string('namechange', 'mod_collaborate'), $namechangeurl);
 }